update reactphysics to v0.8

+++ src/afk/ai/ChunkyTriMesh.cpp
@@ -69,7 +69,7 @@ bool ChunkyTriMesh::init(const float *verts, const int *tris, int ntris,
   return true;
 }
 
-int ChunkyTriMesh::get_chunks_overlapping_rect(float *bmin, float *bmax, int *ids,
+size_t ChunkyTriMesh::get_chunks_overlapping_rect(float *bmin, float *bmax, int *ids,
                                                const int maxIds) const {
   // Traverse tree
   int i = 0;
